+++ src/Home.js
@@ -24,25 +24,43 @@ const Home = () => {
     };
 
     function handleEditorDidMount(editor, monaco) {
-        monaco.editor.defineTheme("onyx", {
+        
+        monaco.editor.defineTheme('vortyxTheme', {
             base: 'vs-dark',
             inherit: true,
             rules: [
-                { token: 'comment', foreground: '#5C6370', fontStyle: 'italic' },
-                { token: 'comment.js', foreground: '#8F93A2', fontStyle: 'bold' },
+                { token: 'tag', foreground: '#e258c0' },
+                { token: 'attribute.name', foreground: '#b5469a' },
+                { token: 'attribute.value', foreground: '#ffffff' },
+                { token: 'string', foreground: '#ffffff' },
+                { token: 'delimiter.angle', foreground: '#dbbcd6' },
+                { token: 'delimiter.curly', foreground: '#dbbcd6' },
+                { token: 'delimiter.square', foreground: '#dbbcd6' },
+                { token: 'keyword', foreground: '#775271' },
+                { token: 'number', foreground: '#775271' },
+                { token: 'meta.tag', foreground: '#775271' },
+                { token: 'meta.tag.custom', foreground: '#775271' },
+                { token: 'meta.tag.doctype', foreground: '#775271' },
+                { token: 'meta.tag.inline', foreground: '#775271' },
+                { token: 'meta.tag.script', foreground: '#775271' },
+                { token: 'meta.tag.style', foreground: '#775271' },
+                { token: 'meta.tag.xml', foreground: '#775271' },
+                { token: 'comment', foreground: '#537178' },
             ],
             colors: {
-                'editor.background': '#282a36',
-                'editorCursor.foreground': '#ffffff',
-                'editor.lineHighlightBackground': '#222432',
-                'editorLineNumber.foreground': '#dbdee1',
-                'editor.foreground': '#dbdee1',
-                'editor.selectionBackground': '#3d465b',
+                'editor.background': '#161c1c',
+                'editorSuggestWidget.background': '#161c1c',
+                'editorHoverWidget.background': '#161c1c',
+                'editorFindWidget.background': '#161c1c',
+                'editorFindWidget.input.background': '#161c1c',
+                'editor.foreground': '#dbbcd6',
+                'editor.cursorColor': '#ffffff',
+                'editor.lineHighlightBackground': '#121717',
+                'editor.selectionBackground': '#1f4953'
             },
         });
-        monaco.editor.setTheme("onyx");
+        monaco.editor.setTheme('vortyxTheme');
     }
-
 
     return (
         <div className="editor-container">
@@ -50,10 +68,10 @@ const Home = () => {
                 <Editor
                     height="100vh"
                     language="html"
-                    theme="onyx"
                     value={code}
                     options={{
                         language: 'html',
+                        fontFamily: 'Consolas !important',
                         minimap: {
                             enabled: false
                         }